move duplicate models to cbp-utils & fix references
