+++ packages/cbioportal-utils/src/index.tsx
@@ -21,11 +21,13 @@ export * from './model/Mutation';
 export * from './model/MyCancerGenome';
 export * from './model/MyVariantInfo';
 export * from './model/OncoKB';
+export * from './model/OncoTreeTumorTypes';
 export * from './model/PostTranslationalModification';
 export * from './model/RemoteData';
 export * from './model/SignalMutation';
 export * from './model/SignalMutationStatus';
 export * from './model/SignalGeneFrequencySummary';
+export * from './model/TherapyRecommendation';
 export * from './model/Uniprot';
 
 export * from './mutation/MutationAnnotator';
