+++ packages/react-mutation-mapper/src/component/mutationMapper/MutationMapper.tsx
@@ -30,7 +30,7 @@ import {
     TrackName,
     TrackVisibility,
 } from '../track/TrackSelector';
-import { ISharedTherapyRecommendationData } from '../bzkfAnnotation/modelCopy/TherapyRecommendation';
+import { ISharedTherapyRecommendationData } from 'cbioportal-utils/src/model/TherapyRecommendation';
 
 export type MutationMapperProps = {
     hugoSymbol?: string;
