+++ packages/react-mutation-mapper/tsconfig.json
@@ -2,5 +2,5 @@
   "extends": "../config/tsconfig.json",
   "include": [
     "src",
-    "../typings/**/*" ]
+    "../typings/**/*", "../cbioportal-utils/src/model/OncoTreeTumorTypes.tsx", "../cbioportal-utils/src/model/TherapyRecommendation.ts" ]
 }
