+++ src/pages/patientView/therapyRecommendation/FollowUpTable.tsx
@@ -8,7 +8,7 @@ import {
     IDeletions,
     IMtb,
     IResponseCriteria,
-} from '../../../shared/model/TherapyRecommendation';
+} from 'cbioportal-utils';
 import { computed, makeObservable, observable } from 'mobx';
 import LazyMobXTable from '../../../shared/components/lazyMobXTable/LazyMobXTable';
 import styles from './style/therapyRecommendation.module.scss';
