+++ src/pages/patientView/therapyRecommendation/TherapyRecommendationTableUtils.tsx
@@ -1,9 +1,9 @@
 import {
     ITherapyRecommendation,
-    EvidenceLevel,
+    TREvidenceLevel,
     IReference,
     EvidenceLevelExtension,
-} from 'shared/model/TherapyRecommendation';
+} from 'cbioportal-utils';
 import _ from 'lodash';
 import request from 'superagent';
 import * as React from 'react';
@@ -38,7 +38,7 @@ export function getNewTherapyRecommendation(
         id: patientId + '_' + timeId,
         comment: [],
         reasoning: {},
-        evidenceLevel: EvidenceLevel.NA,
+        evidenceLevel: TREvidenceLevel.NA,
         evidenceLevelExtension: EvidenceLevelExtension.NA,
         evidenceLevelM3Text: '',
         author: getAuthor(),
@@ -74,7 +74,7 @@ export function isTherapyRecommendationEmpty(
     if (
         therapyRecommendation.comment &&
         therapyRecommendation.comment.every(s => s === '') &&
-        therapyRecommendation.evidenceLevel === EvidenceLevel.NA &&
+        therapyRecommendation.evidenceLevel === TREvidenceLevel.NA &&
         _.isEmpty(therapyRecommendation.reasoning) &&
         therapyRecommendation.treatments.length === 0 &&
         therapyRecommendation.references.length === 0
